Rethink usage of "dependencies" and "devDependencies"
As I learned, by using a bundler and only using this in the frontend,
the distinction as initially used is pretty useless. So we could just
put all packages in either of the two.

We split them this way now, though, as it means that someone just
building the app only needs the non-dev deps. Only "optional" stuff
such as type checking and linting requires dev dependenies. (It hurts
my heart to write that type checking is optional...)