Remove `prod` feature hack and use `debug_assertion` to check for dev/release
We want to have slightly different semantics for our release build that
we deploy (e.g. we are more strict when parsing the configuration and
use fewer defaults). We used a dummy `prod` Cargo feature for that
purpose because we didn't want to complect "optimizations" with the
"target" of a build (for production, for development). However, this
is a bit ugly and there is actually no reason to use optimizations
during development. There are only two reasons to enable optimizations:
deployment and benchmarking. The latter is so rare that the requirement
to have a config file is not too bad.

So that's why the dummy feature is removed in favor of the standard way
to check: `debug_assertions`.

Additionally, that way we match the behavior of the `rust-embed` crate
we will be use in the future.