Add watchboi configuration
Well, this is an experiment. Watchboi is a tool that I wrote last week
for exactly for projects like these. As such, the tool is really not
mature yet. The reasoning was like this:

We want a convenient development workflow/loop. When I change files,
everything should be automatically recompiled or rebundled and my page
in the browser should refresh. Of course, there is webpack-dev-server
which works really well for SPA. But it also works with a backend API
server as it can proxy through all requests to the API! Great!

However, we can't automatically recompile the backend if needed. We
could start a separate watch process, but then the dev-server does not
correctly refreshes the browser tab.

An alternative solution would be to add some of this "development
workflow" logic into our code. For example, our backend could watch for
changes and stuff like that. But ultimately, I don't think it's a good
idea to add logic related to the development tools into your
application. That way, there are only very few code paths that are only
hit in development or only hit in production.

So then we need another external tool. But a tool that works with Rust
and JS, or even better: be completely language-agnostic. I didn't find
a tool with the requirements I had in mind. But I thought that such a
tool would be really useful for many projects! So I wrote one.

Watchboi can do everything we need (so far). In particular, it can
inject auto browser refresh JS code into the HTML returned by the
reverse proxy server. All this means that I can just put it in front of
our backend and it just works. There surely are still bugs and all,
but I used it for the last 2 hours or so and it works fine.