Add Apollo implementation of example
It works, but one thing is still problematic: the `gql` JS template
does not generate TS typings. There are a few way to pre-process gql,
but none of those generates types either I think. We can manually
specify types, but this feels like duplication, as the schema already
defines all the types.

Maybe specifying types in the frontend again is good to make sure
frontend and backend agree on the types. But in that case we would like
to have a test to check that both agree on the types. So in either way,
we need some way to generate types from the schema and `gql`
invocations.