Configure `browserslist` and `@babel/preset-env`
By default, and mostly for backwards-compatibility-reasons,
the `env` preset behaves like the now deprecated `latest` preset,
which basically means everything will be compiled to ES5.

However, this level of compatibility is overkill for most apps,
and to circumvent it, we have to be explicit about which browsers
we want to support.

This is done in the `"browserslist"` key in `package.json`
as recommended by the `browserslist` documentation.
This configuration should then just be automatically used
by all the tooling using `browserslist` internally.
However, unfortunately there is a [bug](https://github.com/babel/babel/issues/10867)
in Babel that makes it ignore that setting
when it contains the recommended `defaults` query.
The workaround for that is to get even more explicit
and tell Babel about the config manually.