Add emotion as CSS library
I am not sure that emotion is the best CSS library for us, but I guess
it will work fine, at least for now. In particular, emotion seems to be
the only (popular) library that allows to specify CSS properties
directly on JSX elements via `css={{ ... }}` which is what I want, I
think. It also works fairly well with Typescript, I think.

This installation of the package in particular:
- The JSX factory method is set globally for Babel and TS, so no
  `/** @jsx jsx */` annotations are required in each file.
- Unfortunately, I didn't manage to make tsc automatically add an
  import for `jsx`, so now `import { jsx } from "@emotion/core";` is
  required in all files using JSX.
- This already uses the emotion babel plugin which improves several
  things about the generated code and classnames.