Add "Open Sans" font files and add `@font-face` CSS declarations
The typeface is licensed under Apache-2.0. This permits to bundle those
files with our application, as far as I can tell. It has been split
into multiple files that each only contain a subset of glyphs.

I decided to put all the `@font-face` declarations directly into
`index.html` to avoid another HTTP request. It is a lot of text (10230
bytes), but gzip on level 2 compresses it down to only 755 bytes, which
is not a problem.