Disable "react/prop-types" lint
For now we decided that we don't need prop types. Typescript and the
well typed API will already catch tons of errors. If we encounter a
situation where prop types are useful, we can still add them. And if we
find them so useful as to enforce them everywhere, we can do that
later still.