Remove assets startup check
This currently leads to `floof` failing on a clean build if the
backend is started before the frontend was built. In either way, it
shouldn't run in debug mode. But since this check is flawed anyway and
since I will change the way we load assets completely in the future, we
just remove it for now.