Add `util` crate with convenience `prelude`
Some symbols are used in almost every module and are annoying and noise
to import. Especially the log macros: you often want to quickly, maybe
temporarily add a `debug!` output. Adding and then removing the import
is very distracting. This prelude just reexports all of those always
important symbols.

The `util` crate will surely be useful for other stuff in the future.