Install an error boundary around components talking to the backend
This handles both, the new `APIError` we throw for >=400 requests,
and the case where the API returned **no** `data`, only `errors`.
Note that GraphQLs ability to model partial responses with errors
does not fit well into this framework. You will get the partial data,
but there is no way to get to the errors in that case, using Relay's
hook API. Specifically, the error boundary does **not** handle that case.

Note also that this boundary will catch every other error, apart from the
[usual exceptions](https://reactjs.org/docs/error-boundaries.html#introducing-error-boundaries),
that is thrown below it. If we want more granular error handling down
the line, we will need multiple, potentially nested error boundaries
inspecting their errors a bit more closely than this example.