Add mechanism to deal with DB migrations
My solutions looks like this: maintain a table `__db_migrations` which
contains a row for each active/applied migration. It contains its id,
name and the raw SQL script that was used. The latter is important for
debugging later on, also in production. It is also used to make sure
that the applied migrations fit the migrations the application knows
about; otherwise changing a migration is silently ignored.

The `db::migrate` function automatically creates `__db_migrations` if
it doesn't exist yet. It then checks everything it can and runs all
missing migrations. The actual migrations scripts are defined in
`src/db/migrations/`.