Add `tobira db console` convenience command
This gives you a `psql` prompt for the configured database.
It replaces the `tobira` process so that signals, streams, etc. are
handled correctly, and so that we don't have to care about errors.

Note that this is a Unix-only feature. Note also that this is like
calling `process::exit` in that it does not run any destructors.
The internal `execvp` call guarantees that all memory is freed
correctly at least, though.