Switch from base85 IDs to base64 IDs
The original idea behind using the less common base85 was to save space.
We have 64 bit keys in the database which can be encoded with 10 bytes
via base85. On the other hand, base64 would need 11 bytes to do the
same. But we noticed that we need URL-safe IDs anyway and we don't want
to use two different encodings for our IDs.