Change prefix of "series" to "sr"
Well, this is awkward. But we noticed that having "se" as prefix for
our stringified IDs leads to a problem. With base64, we have a 2/64
chance of any digit, including the first one, being an "x" or "X".
(base85 is not particularly better in that regard.) While in URLs, we
will probably omit the prefix (as it's implied by `server/series/123`),
it is very likely users will see the stringified IDs with prefix at
some point. So simply to prevent the potential future stupid situation
of some lecturer complaining that their series URL contains "sex", we
change the prefix.