Slightly refactor API error handling in the frontend
Unfortunately, Relay modern makes this a bit awkward.
Normally, if your GraphQL response contains any `data`,
your components just gets that as result from the Relay hook.
Even if there are `errors` in the response, as well!
Only if there is **no** `data`, and **only** `errors`
would you get an exception from somewhere deep inside Relay,
which you then had to parse out in an error boundary.
(See the previous version of our `APIErrorBoundary`
for an example.)

That means that you can either get the data, but not the errors,
or you can get just the errors in a very weird form.

What we do now is inspect the response ourselves,
before giving it to Relay, and throw a custom error
if it contains any `errors`.
This way, you can still never get the `errors` in your component,
but at least you have both, `errors`
and the potentially partial `data`
in your error boundary in case of an error.

As a result, if you want to handle API errors in your component,
you will probably want to model explicitly as data,
as opposed to relying on `errors`.
Otherwise you would have to extract the rendering logic
from the GraphQL query,
so that you could then call that rendering component
from the error boundary, giving it the partial data.
This would just be a poor person's reimplementation
of Relay Classic's HoC-s, eliminating the elegance we get
from using Relay Modern's hooks.