Fix dynamic base path for assets
The problem is that the basepath set in `assets!` is relative to
`CARGO_MANIFEST_DIR` at compile time, but relative to the current
directory at runtime. Since the latter can vary, we introduce a config
value which points to the correct folder. Since that path is relative
to the config file, that works out, regardless of where the backend
is started.