Add `serve` subcommand to replace subcommand-less invocation
Just starting the binary without arguments now results in the help text
and does not start the backend HTTP server anymore. This is done with
`./tobira serve`. This is to make starting the server more deliberate.
We have quite a few subcommands now, so having a subcommand for the
server makes for a clearer CLI.