Render text blocks as Markdown (a subset of CommonMark)
This uses the package `react-markdown`. Unfortunately, this adds 32kB
to our gzipped bundle size bringing it from 174kB to 206kB. I only
found one other library that could be used and was only 5kB. However,
it was not as popular and more importantly: it didn't have as good of
a default behavior regarding raw HTML and didn't allow to customize
which elements are allowed.

The markdown that's being rendered here is a subset of CommonMark. I
did not want to allow everything right away for several reasons. Most
notably missing: headings and images. See code comments for more
information.

Unfortunately also missing are autolinks and strike-through. The former
is something I am used to with GitHub flavored Markdown. Just writing
`github.com` or `https://foo.de` would create a link. Not so in this
PR, where everything has to be explicitly linked with `[]()` or
"explicitly autolinked" with `<https://foo.de>`.

Raw HTML is escaped.