Use different ID for deployments (used in e.g. the subdomain)
The previous ID always contained a base64 string, making it hard to
read and type by hand. It also didn't consider PRs. With this commit,
PRs have the id `pr<id>` and branches simply use their branch. However,
if the branch is truncated after 40 chars and any special characters
are removed. In those cases, a short hash is also appended to guarantee
uniqueness.