Add test deployments for PRs by splitting workflow
Our goal is to test deploy pull request so that one can easily test the
changes without checking them out locally. The problem is that we need
secrets for the deployment (i.e. to access the server). Just running
the script with secrets for random PRs is a security problem as CI
essentially executes arbitrary attackers code.

The solution, which I luckily found in a blog post, is to use the
`workflow_run` trigger to start one workflow after another one has
finished. The first workflow builds and tests the actual project code
and has no access to secrets. Some of those results (e.g. the resulting
binary) are then archived as artifacts.

The next workflow can then download those artifacts and do the
deployment. That second, `workflow_run`, workflow does have access to
the secrets.

However, we still essentially take one executable that is fully
controlled by the PR author and execute it on our servers. Here is
where the last point of defense comes in: we only deploy if
`github.actor` is one of three blessed people. But: this check is
defined in the `deploy.yml` file, so what if an attacker just edits
that to remove the chat? Well, `workflow_run` triggered workflows run
in the context of the main repo and always use the version on the
default branch. So changes in `deploy.yml` are not actually used until
the PR is merged (and well, this shouldn't happen if the change is
malicious).

Somewhat annoyingly, this means that to test my script, we actually
need to merge my PR now. No way to test it before.

And finally: at some point we might want to also test deploy PRs from
other people by first doing a code review and then commenting `/deploy`
on the PR or something like that.