Remove old menu code and language icon
The old menu is likely to be replaced fully by the new burger. I liked
the language icon and think there is some use in making the language
switcher quickly and easily available. However, others convinced me
that it's mostly useless. The language is already guessed from the
browser (and that's rarely incorrect; and if it it it's the user's
fault). And Tobira will not support content in multiple languages for
quite some time, so the language switcher is not really that useful
compared to, for example, Opencast Studio. The latter application has
lots more "application text" and a lot less "content text".

Of course, the ability to switch the language will be added back at
some point.