Start loading query in route `prepare` function, not during render
This uses new relay hooks, namely `loadQuery` and `usePreloadedQuery`.
Both are intended to avoid multiple nested or waterfalling round trips
and to speed up the UI by starting to fetch as soon as possible.

The docs recommend "prefer calling `useQueryLoader` when possible". I'm
not sure but I think we cannot use that. All examples using it seem so
far removed from what we need/have here. In our application, we
basically always change the route when a new request to the API is
necessary. So yeah, seems unfitting for us.

Unfortunately, that means we have to somehow deal with disposing of the
query ref that `loadQuery` returns. Otherwise it leaks memory,
apparently. This will done in a later commit.