Update juniper to tokio 1.0 compatible version (not upstreamed yet)
I got really tired of waiting and see random Tokio runtime errors due
to us having two Tokios runtimes. There is an open PR for a long time
already to update juniper accordingly. It has not been merged yet. But
apparently people have already been actively using it for a long time.
And everything appears to work here.

We probably want to switch away from Juniper entirely soon, but this
was a fast update and gets rid of some annoying things, so here we go.