Update to React 18 alpha
We want this for multiple cool new features, like `startTransition`.
This is already used in this commit and avoids a loading spinner
flashing in between requests. Of course the current solution is not
perfect yet because there is no loading indicator at all. But we can
fix that in the future.

Of course, using alpha software is kind of meh. React has been pretty
stable though. We know we will switch to React 18 at some point, thus
we might already use its benefits now. I don't expect updating to beta,
RC and then stable to be problematic.

Kind of unfortunately, `@types/react` does not publish an alpha version.
However, the correct typings are already hidden inside of it. To use it,
one can use different methods, but what's implemented in this commit
is the best I was able to achieve.