Change routing (remove `/r/` prefix, use `~` for internal paths)
This was planned for quite some time. We wanted to get rid of the /r/
prefix for realm paths. The new rules are roughly like this:

- `/~` -> internal/static paths. This is for all non-user provided
  content. In particular, `/~assets/` is the new "directory" for all
  static assets.
- `/foo/bar` -> realm pages.
- `/foo/bar/v/<video-id>` -> video pages. Videos directly belonging to
  the root realm will have an `/v/<video-id>` path, i.e. what they had
  before.
- `/@userid` -> for later, this is the personal page of users.

Right now, there are still two problematic routes: `/graphql` and
`/graphiql`, both of which could clash with realm routes. For `graphql`
it's potentially not problematic, because we only allow POST requests
to that, but `graphiql` is certainly problematic. We can probably just
use `/~graphiql`.