Make navigation usable with query or static data & fix video route nav
Before, the navigation had to be given fetched data. This was bad in a
couple of ways, in particular also leading to duplication in root
queries. Now the navigation components can be used with static data or
with a relay fragment reference. The machinery necessary is a bit
involved but it's still fine.

This enabled us to show proper navigation on video pages.