Silence emotion-js warnings about `:first-child`
We don't do SSR and we are currently not planning on doing it. Having
those warnings pop up all the time and having to rewrite `:first-child`
selectors is just annoying and not feasible. This is sadly the best
way to globally silence those warnings. While this adds a new
dependency in `package.json`, it doesn't add to the bundle size since
it was in the dependency graph already.