Fix check in deploy GH action
Oh my lordy. That single character wasted way too much of my time. With
it, the multi line string literal has *no* trailing newline. Without it,
there is a trailing newline. Which means the value of `if` was
`${{ ... }}\n`. The thing inside `${{ }}` was evaluated to `false`
resulting in `false\n`. And since that's a non-empty string, it is
truthy I guess. Weak typing strikes again! Amazing.

Fixes #231