Disable eslint cache
I saw several repeated problems with it, unfortunately. Lints not being
reported or being reported despite the file having changed. With the
new workflow of running eslint in `webpack watch`, it does not seem to
make any difference in speed. So we don't really benefit from the
cache, just get confused by the anomalies it causes.