Use one DB transaction per API request
This has a number of advantages.
- By checking out one connection from the pool at the start of the
  API request instead of in every single handler, we avoid useless
  work.
- Using a transaction means that each individual API request gets a
  consistent view of the world.

Unfortunately, we cannot achieve this without `unsafe` code. Juniper
does not support contexts with lifetime parameters (but `Transaction`
has a lifetime parameter). While we would love to get away from juniper
for a couple of reasons, we don't have the time to do that. So we
cast it to a `'static` lifetime temporarily. We make sure that it does
not outlive its connection. So I'm fairly sure this all is sound.

There are still a couple of things we should do:
- Currently, the full realm tree is loaded on each API request. That's
  wasteful. I will solves this completely differently in a later commit.
- We should think about what isolation level our transaction needs and
  specify that explicitly.
- We should mark the transaction as "read only" if there are no
  mutations in the GraphQL query.
- We should retry after a failed transaction commit, at least a couple
  of times.

The last two points are impossible without coding the `juniper_hyper`
logic ourselves.