Store full realm path instead of path segment per realm
Turns out that index expressions need to be "immutable", i.e. pure
functions. Bummer. But if we could have used `full_realm_path` there,
that would be almost too good to be true. So instead we store the full
realm path in a normal column.

This makes editing a path segment a lot more annoying, but we need some
index of full path to realm since that lookup happens on virtually
every page view. So the computational work is required anyway. If we
store the full path anyway, we don't need to store the last segment
additionally.

One alternative would be to have a separate table to serve as index and
have triggers that modify that table on path_segment changes. That way,
we could edit path segments via GUI DB tools.