Switch from custom config macro to `confique`
This was planned for a long time already. In fact, our own macro in
this code base was written to be easily exacted into a separate lib.
However, the initial idea was somewhat flawed and I now think that a
`derive` based macro and library is better for several reasons. That's
what `confique` is.

One immediate advantage is that now, different parts of the config type
can be defined in different modules. This is what this commit already
does which is why it touches so many files.

Apart from just switching to the new lib, I changed the following
things on the way:

- `db.user` has a default now: "tobira"
- `db.host` has a default now: "127.0.0.1"
- `opencast` was renamed to `sync`
  - `sync_user` is now `user`
  - `sync_password` is now `password`

With this commit, Tobira does not fully use all features of `confique`,
mainly, layered configuration is not used. That would allow us to
load and merge config values from multiple sources. We can add this
later.