Add `child_order` column to realms and do some sorting in the frontend
Before this, Postgres did the sorting by name. But this is not perfect
because we are fixed to Postgres' way of sorting strings (pretty much)
and the sorting is not locale aware. There were some place in the
frontend where we want to sort in the frontend anyway, just to show
a preview, for example. Finally, we might want to be able to sort in
a way other than just "alphabetic ascending".

So now there is an explicit field on each realm saying how to sort its
children. The backend returns children always sorted by index. The
frontend is supposed to sort differently if `childOrder` says so.