Strip down landing "manage" page to a simple "TODO"
We noticed that we don't fully know how we want this to be. So we are
waiting for more "manage"-like pages to be implemented. Further, we
want to get rid of the navigation "dispatch" stuff: we don't need
static navigation data. That can simply be created as separate box,
like the extra manage box on realm pages.