Remove static/dynamic data source dispatch from `Navigation`
I first thought we needed this because we want some static navigation
for things like the management area. Another reason was that the nav
was different for the burger and desktop version, so it made sense to
just pass the data source to `Root` and let it create two different
nav elements from it.

But after adding the box with "realm settings" and "add sub page"
buttons, I realized that static navigation does not need to go through
the `Navigation` component. One can just pass `JSX.Element` to `Root`.
Well, but there is still the burger/desktop divide. It would still be
fine if components just pass both versions to `Root`.

But luckily we also made the burger/desktop navs almost the same now.
With this, we can remove a lot of code and let `Navigation` just be
about realm navigation.