Merge all three backend crates into one (major refactor)
We moved to a multi-crate model at some point for two reasons: (a) we
had a proc-macro crate which needs cannot be in the main crate, and (b)
we had some circular dependencies between the GraphQL schema (required
by frontend, exported by backend) and the frontend artifacts (required
by the backend).

Reason (a) has disappeared since transitioning to `confique`. Reason (b)
also does not really apply any more: the GraphQL schema is stored in
the repository and thus rarely needs to be generated from Rust code
anyway. Also, the frontend artifacts are not required to build the
backend in debug mode. So there are lots of escape hatches from this
circular dependency.

The advantages of a single crate is simplicity. It has become apparent
that the `api` crate really didn't make a lot of sense as an own crate.
There are a lot of dependencies between the API code and the rest of
the code. So the separation was artificial anyway. This also resulted
in a fairly useless `tobira-util` crate.