Change API to return `"/"` as path of the root realm
This makes sense in more places in code and fixes a bug where one would
not be forwarded to the root realm when deleting a child of the root
realm (because it was forwarded to relative URL `""`). Both view points
make sense, so we just take the one thats more convenient to us.