Restructure translations of error messages
I had enough: I settled for the slightly less optimal format:

    Action failed. Reason for failure.

instead of the slightly superior:

    Action failed: reason for failure.

The reason is that the reason strings are used in different contexts,
sometimes as a sentence on its own. So there they would have to be
capitalized. Sure, there are some formatting filters and all that jazz
that you can do. But it makes stuff a lot harder. So now I decided to
just concat the two sentences. It's fine. Users will still understand
what we mean.