Transmit theme information via CSS vars instead of JSON data
Before, theme data was transmitted via JSON in `index.html` and the
frontend had to parse the colors, create variations of them and store
all of that in CSS vars. With this commit, the backend does the color
parsing, creating variations and setting them as CSS vars inside the
`index.html`. This cuts down on JS code in the frontend.

It might seem a bit odd at first to have such "style related" things
in the backend, but it's fine. The backend already defines the color
config, so just processing these values in the backend is fine.