Add query parameter support to route and use it across routes
If each route has to check for required query parameters manually, that
adds a lot of useless dummy code. It also makes it especially difficult
because the GraphQL query was often optional. But we want to show user
information on each page, which always requires a GraphQL query. So
that would have resulted in having two different queries or using dummy
query parameter. All of that is not great!

This way, this can be handled in a central fashion. Routes just declare
which query parameters are required. If those don't exist, the route is
not taken. If they are, they are passed to the route. All of that in a
fairly type safe fashion.

You might think: but is showing the 404 page good UX if the parameter
is missing? For one, that should never happen and is always a bug, and
additionally, the error pages we have shown before are also not really
better.

Finally, a note about implementation. There are some suboptimal things
about this.

The `queryParams` field always have to be specified
(leading to a few `queryParams: []` additions). If it was optional, one
could specify the type parameter `QueryParams` to something non `[]`,
then access those parameters in `prepare`. But since `queryParams` (the
runtime value that the routing code needs to decide stuff!) is not
specified, no query parameters are checked and passed to `prepare`,
leading to surprising bugs.

One idea was to only have the field if `QueryParams != []`. I tried
something like this:

    type Route<Prepared, QueryParams> = { ... }
        & (QueryParams extends [] ? {} : { queryParams: QueryParams });

But that showed another TypeScript unsoundness: a function generic over
`Route<P, Q>` could always access the field `queryParams`. Despite it
not existing if `Q == []`.

Further, I did not find a way to not specify the list of parameters
twice. As you can see in the routes using parameters, they are specified
as field and also as type parameter of `makeRoute`. TS cannot derive
runtime values from generic parameters as everything is just type
erased. The other way around should work, but here, TS shows some weird
type inference behavior. Consider this:

    function foo<T extends string[]>(t: Readonly<T>): Record<T[number], string> | null {
        return null;
    }

    const nope1 = foo(["hi"]);
    const nope2 = foo(["hi"] as const);
    const works = foo<["hi"]>(["hi"]);

`nope1` and `nope2` have the type `Record<string, string>` as `T` in
those invocations is `string[]` instead of the tuple type `["hi"]`. This
is super weird! Especially since `as const` was especially designed for
this stuff.

Either way, specifying it twice is annoying but at least the two copies
cannot disagree since it's type checked.