Remove circular dependency checker plugin
This was added in 3772213b4b194272aa898157c104aa27331c6fa5 because I
noticed one weird bug that was actually caused by circular dependencies.
However, as that commit already said, they mostly do not cause problems.
That particular problem was caused because of some global const values
depending on one another.

All current circular dependencies are basically just

    "some-route" needs `<Link>` from `router`
    `router` needs route description from "some-route"

But that's not problematic: the `<Link>` is only used within functions.
And with the new router design, the only way to break those dependencies
is split all routes' component and route description into two files.
And that's not great I'd say.