Refactor handling of session cookie `MaxAge`
This moves the conversion from `std::time::Duration`
to `time::Duration` to the `SessionId` type,
and takes some of the love and care out of the error handling
for some easier to read code.

To make up for that it also renames the `set_cookie` parameter
in order to clarify that we expect a session duration, as parsed
for example from our config, instead of some generic `Duration`.