First draft of block deletion API
Note that it returns the realm the block was deleted from as well.
This makes it easier (automatic, in fact) to redraw any listing
of blocks that you delete from in the frontend.
For similar reasons, it also does not treat deletion of blocks
that don't exist (anymore) as an error.