Work around some Juniper problems
Juniper is currently not generating valid schemas for interfaces,
at least the way we set them up here. Every implementation
of an interface has to duplicate the fields of the interface
in the schema. Juniper doesn't do this atm, so we have to
help it along.

Probably related to this, Juniper also sometimes crashes when you
put a fragment on both, the generic interface of a field, and its
possible implementations. Juniper is then looking for the
generic fields in the derived objects, and presumably can't find them.
This is also fixed by this workaround.