Add first usable version of the uploader
Quite a large commit, moving some code around, improving lots of UI
and ultimately made the main uploader flow working. Quite a few things
are still missing or problematic. Most notably: error reporting is
pretty bad right now. And: the JWT only gets evaluated once the
`addTrack` is done. Which is problematic since we might not want to
have the JWTs be valid for that long.