Add first version of pagination for `myVideos` endpoint
Where to begin? The Relay pagination specification is pretty bad IMO.
Apart from bad, confusing and incorrect wording, there are a few design
flaws I think.

For one, `PageInfo` is defined as a global type but I
think it makes a lot of sense to have different per-page information
for different connections. In this case, I want to return the indices
of the first and last item in the list. Also, the `startCursor` and
`endCursor` are declared as non-null, which makes no sense. Finally,
the "edge" object makes little sense I think. It only provides the
cursor, but in all situations I can imagine, you are only interested in
the start and end cursor.

Regarding implementation: it's not trivial and quite a bit of code. The
code I ended up with is fairly fine I think. It covers a few cases that
are rarely (if ever) useful, but that wasn't a lot of extra work. I hope
this can somehow be reused for other connections in the future.
Otherwise... uhg. We will see how we will design connections in the
future, I guess.

The frontend does not yet really use the pagination capabilities. Its
query was adjusted to work for the first 15 events, but it does not
show any pagination UI.