Remove `EventEdge` and return `items` directly in `EventConnection`
The cursor inside the `Edge` is not used by any frontend code and is
just from old relay. And if there is no `cursor` in `Edge`, there is
really not any point in keeping the edges: we can return the items
directly.

Of course, we stray further from the relay connection spec and can't
use some relay functions then. BUT for the "my videos" page, we were
not able to use them anyway. So we might as well have a proper GraphQL
API for this endpoint.