Improve scrolling behavior on navigation
Before, clicking a link to a new route would keep the old scroll
position, which is weird. With this commit, each click on navigation
links will scroll to the top.

Before, the browser did not restore the scroll position when clicking
"back" in the browser. With this commit, it does.

However, this solution is not perfect! On navigation click, the page
is instantly scrolled to the top, potentially before the new page has
loaded. This is weird as well, especially so for very slow loading
pages.

Also, when going back or forwards, the automatic scroll restoration of
the browser only kicks in after between 30ms and 100ms, meaning a few
frames are still rendered with some wrong scroll position. And we also
have no coordination with when the previous page is actually loaded.

So I think the only proper solution involves scrolling manually when
the navigation `startTransition` is done. And to access scroll
positions, we likely want to store them in the history stack by using
`replaceState` whenever scroll position changes or on navigation.