Improve scroll restoration and scroll behavior during navigation
I am actually very happy with the result and don't see what could be
improved upon it. Instead of letting the browser do scroll restoration,
we do it manually to avoid the disadvantages of the former approach
(see last commit for more information).

Two main ideas behind this approach:

- We store the scroll position of each history entry in its state.
  That's better than having some external hash map keyed with URL. To
  maintain the correct scroll position, we update it for each `scroll`
  event via `history.replaceState`.

- We want to only set the scroll position via `window.scrollTo` once
  the route is actually rendered! Otherwise it looks jarring and
  confusing. To do that we have to wait for a React transition to
  finish. There is no direct way to hook into that, but we can get
  there by using `useEffect` that is only executed when the active
  route changes.

With that, all problems mentioned in the last commit should be solved
by this.