Fix creating multiple series blocks in one session
The dummy series in the Relay store was created anew every time
which of course errors since we reuse the same ID.
We now create it lazily only if it's not already there.
This should also work in the case where Relay's GC
might have picked it up already between two series block creations.