Improve "empty page" and "no videos" cases for "my videos"
The "no videos" page now got a simple note. Maybe we want to extend
that in the future. Something like "how about uploading or recording a
video?".

The "empty page" can never happen with a static data store. But if
someone loads a page and gets `hasNextPage: true`, but then someone
else deletes the only video that would have been shown on that next
page, then that page is empty. We don't get proper indices for the
page, so `?` are shown instead. The returned `has{Next|Previous}Page`
values are useful though! So the user can always get back to the
previous or first page.