Return `Option<&UserData>` from `current_user` and remove manual impl
It would be nice if `User` (which represents a potentially logged-in
user) would implement `GraphQLValue` directly as `Option<UserData>`
would. This is what happened before this commit. Sadly, implementing
it that way requires a lot of boilerplate code that is likely to get
slightly buggy. So instead, we just use `Option<UserData>` directly
to get rid of that code. Slightly less nice, but works.