Update relay to v13
I needed to specify a folder for schema extensions. I haven't found
this in the release notes or anything, just found that fix by trial
and error.

NOTE: the new relay-compiler requires a fairly recent version of
`watchman`! If you are getting weird watchman errors: that's it.

Also, the artifacts directory has to already exist for the compiler to
work. Which is kinda meh. But the easiest is to just check the
directory into git.